Add game grid, add grid and square styling
This commit adds the basic functionality for the etch a sketch project.
To start this project three essential functions were added
toggleSquare(), createGrid() and addSquareListeners(). In styles.css
flexbox was used to create the grid of squares. Each square is set to
grow to remaining space which will allow us to increase or decrease the
amount of squares and fill the #etchGrid container.