Add grid size, clear all button and add their functions
This commit adds two buttons grid size and clear all to the etch
a sketch. Clear all is a clickable button that allows the user to remove
all tiles with the .squareEnabled class. This will completely clear
the grid and allow the user to start from scratch. The change grid size
button will ask for a user input on the grid size. The changeGridSize()
function will remove all the rows of squares from the old grid and then
call the same functions as used when loading the default 16 by 16 grid.