Rework clearGrid() add colorful pen toggle
This commit reworks clearGrid() using code from changeGridSize().
This allows us to change the background color of squares dynamically
instead of adding the squareEnabled class to divs. If the colorful
pen is toggled on by the user a random number is generated then the
background color is changed randomly to pre-selected colors.